refactor: tsxファイルの持ち方を変更
